+++ src/main/java/kr/codesqaud/cafe/account/JoinFormValidator.java
@@ -1,6 +1,6 @@
 package kr.codesqaud.cafe.account;
 
-import kr.codesqaud.cafe.account.form.JoinForm;
+import kr.codesqaud.cafe.account.dto.JoinForm;
 import org.springframework.stereotype.Component;
 import org.springframework.validation.Errors;
 import org.springframework.validation.Validator;
