+++ src/main/java/kr/codesqaud/cafe/account/dto/LoginForm.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.account.form;
+package kr.codesqaud.cafe.account.dto;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotEmpty;
