+++ src/main/java/kr/codesqaud/cafe/post/MainController.java
@@ -1,6 +1,6 @@
 package kr.codesqaud.cafe.post;
 
-import kr.codesqaud.cafe.post.form.SimplePostForm;
+import kr.codesqaud.cafe.post.dto.SimplePostForm;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
