+++ src/test/java/kr/codesqaud/cafe/account/UserControllerTest.java
@@ -16,7 +16,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
 
-import kr.codesqaud.cafe.account.form.JoinForm;
+import kr.codesqaud.cafe.account.dto.JoinForm;
 
 @AutoConfigureMockMvc
 @SpringBootTest
